g2d: Add support for setting blitting direction
This is important when handling overlapping windows, scrolling, etc.

Note: the hflip and vflip functionality was previously implemented by
modifying the source direction register, and relied on having the
destination register untouched. When adding blitting direction we have
to modify both, thus had to change the scheme on how we perform
flipping.