ccanyscrn: Introduce struct anyscreen for storing driver context
Using private driver structs is mandatory for storing driver context,
state, etc. We currently only use this to store the miscdevice- and
device instance, but will later extend it to include all the global
variables.

Note: We also introduced one ugly global (anyscreen_global). This was
necessary since we have no access to it otherwise from the exported
function, if keeping the function signature as-is.

Change-Id: Idd46a74c02d26297fe164cb8ae0e762c6f4fa10e