ccanyscrn: Fix driver crash due to using wrong pointer
In the drivers file operations, we've erroneously used the miscdevice
pointer as a struct anyscreen pointer. This was introduced in commit
c6d8488a0d95c0288cd77a318577ad3fbf609275, but wasn't triggered there
since DEBUG was not set, and the anyscreen pointer wasn't used for
anything but logging there. The problem occured when starting to use
members of the anyscreen struct in the later commits.

Change-Id: Iff0fddd9528cfc3df18a2bd84748ba65a1f133d6