ccanyscrn: Add mutex protection during abort and introduce a timeout
fxi_request is called from another kernel thread than the user context
so we have a critical section during the abort mechanism and it needs
protection.

We have no control on when fxi_request is being called by the gadget
driver. We've previously depended on a second call to fxi_request
after an abort, but this does not always happen. Therefore, we've
added a 10 jiffie timeout so that the driver can complete its close
call without receiving the shutdown signal.

Change-Id: I5be7284c4e070923d1ff04cff12e595feb8b2b91