Adds the Realtek rtl8191su driver
This is copied as-is from Hardkernels Odroid tree available at
https://github.com/hardkernel/linux. The driver is located in
drivers/net/wireless/rtl8191su, commit
249d773c8d77e8dcb622de9d4f951653c4897d5a

We add this driver since we have a bunch of Asus N10s at the
office. It is using the Realtek rtl8188SU chip, and there's no
compatible mainline driver for it yet. We've benchmarked the wip r92
driver (http://github.com/chunkeey/rtl8192su), and the r8712u (staging
driver), and this one. The latter is currently the best performing and
most stable of all three.

Change-Id: I319770eaab70ca614991b1e2431b47b1f2985d92